Add files via upload
This is a small mock-up of a football club website i made, significantly better than my previous site as it shows progress in my skills as a web developer. I decided to go simple with the aesthetic, though the colors clash a lot. The next one will be an improvement overall on the visual aesthetic.